capstone_i/requirements.txt
streamlit
openai
python-dotenv
pandas
altair
SQLAlchemy
PyMySQL